nemo: Implement dedicate Theme class
The Theme class handles all interactions with icon themes. In
particular, it handle look-ups for colored versions of folder icons.
This look-up supports theme inheritance: it will look for the closest
ancestor of a theme which defines a folder icon for a directory (while
supporting special directories); then, it will try to find differently
colored versions of that theme.

This is, for example, necessary to properly support Mint-X-Dark, as
there are no colored versions of this theme.